refactor: split revenue and susds modules
