+++ ui-dashboard/src/lib/canonical-revenue.ts
@@ -1,859 +1,38 @@
-import { weiToUsd } from "@/lib/format";
 import { buildDailyFeeSeries } from "@/lib/revenue";
-import { SECONDS_PER_DAY } from "@/lib/time-series";
+import {
+  buildActualAvailability,
+  buildPartialReasons,
+} from "./canonical-revenue/actuals";
+import {
+  buildRevenueBuckets,
+  buildDailySeries,
+} from "./canonical-revenue/daily-series";
+import {
+  buildForecastSource,
+  buildForecasts,
+} from "./canonical-revenue/forecasts";
+import { buildPeriods } from "./canonical-revenue/periods";
+import { buildStreams } from "./canonical-revenue/streams";
 import type {
-  CdpBorrowingFeeSeriesPoint,
-  CdpBorrowingRevenueMarket,
-} from "@/lib/cdp-borrowing-revenue";
-import type { ReserveYieldResponse } from "@/lib/reserve-yield";
-import type { NetworkData } from "@/lib/fetch-all-networks";
+  BuildCanonicalRevenueArgs,
+  CanonicalRevenueResult,
+} from "./canonical-revenue/types";
 
-export const V3_REVENUE_LAUNCH_TIMESTAMP = 1_772_496_000;
-export const V3_REVENUE_LAUNCH_LABEL = "Mar 3, 2026";
-
-export type SusdsYieldDailySnapshotRow = {
-  id: string;
-  chainId: number;
-  token: string;
-  timestamp: string;
-  currentShares: string;
-  costBasisUsdWei: string;
-  realizedYieldUsdWei: string;
-  transferredOutYieldUsdWei: string;
-  redeemedYieldUsdWei: string;
-  currentValueUsdWei: string;
-  unrealizedYieldUsdWei: string;
-  totalEarnedYieldUsdWei: string;
-  dailyEarnedYieldUsdWei: string;
-  dailyRealizedYieldUsdWei: string;
-  dailyUnrealizedYieldUsdWei: string;
-  sharePriceUsdWei: string;
-  sampledAtBlock: string;
-  sampledAtTimestamp: string;
-};
-
-type ActualRevenueValue = number | null;
-
-type ActualRevenueAvailability = {
-  reserve: boolean;
-  reserveStaleAfter: number | null;
-  swap: boolean;
-  cdp: boolean;
-};
-
-export type CanonicalRevenueDailyPoint = {
-  timestamp: number;
-  reserveYieldUsd: ActualRevenueValue;
-  swapFeesUsd: ActualRevenueValue;
-  cdpBorrowingUsd: ActualRevenueValue;
-  totalRevenueUsd: ActualRevenueValue;
-  availableRevenueUsd: number;
-};
-
-export type RevenuePeriodKey = "allTimeSinceV3" | "ytd" | "last30d" | "last7d";
-
-export type RevenueForecastKey = "next7d" | "next30d" | "next365d";
-
-export type CanonicalRevenuePeriod = {
-  key: RevenuePeriodKey;
-  title: string;
-  subtitle: string;
-  from: number;
-  to: number;
-  totalUsd: ActualRevenueValue;
-  availableTotalUsd: number;
-  reserveYieldUsd: ActualRevenueValue;
-  swapFeesUsd: ActualRevenueValue;
-  cdpBorrowingUsd: ActualRevenueValue;
-  partialReasons: string[];
-};
-
-export type CanonicalRevenueForecast = {
-  key: RevenueForecastKey;
-  title: string;
-  subtitle: string;
-  days: number;
-  totalUsd: number | null;
-  reserveYieldUsd: number | null;
-  swapFeesUsd: number | null;
-  cdpBorrowingUsd: number | null;
-  partialReasons: string[];
-  assumption: string;
-};
-
-export type CanonicalRevenueStream = {
-  key: "reserve" | "swap" | "cdp";
-  title: string;
-  actualUsd: ActualRevenueValue;
-  forecast30dUsd: number | null;
-  forecast365dUsd: number | null;
-  subtitle: string;
-  actualPartialReasons: string[];
-  forecastPartialReasons: string[];
-  partialReasons: string[];
-};
-
-export type CanonicalRevenueResult = {
-  periods: Record<RevenuePeriodKey, CanonicalRevenuePeriod>;
-  forecasts: Record<RevenueForecastKey, CanonicalRevenueForecast>;
-  dailySeries: CanonicalRevenueDailyPoint[];
-  streams: Record<CanonicalRevenueStream["key"], CanonicalRevenueStream>;
-  partialReasons: string[];
-};
-
-export type BuildCanonicalRevenueArgs = {
-  networkData: ReadonlyArray<NetworkData>;
-  cdpDailySeries: ReadonlyArray<CdpBorrowingFeeSeriesPoint>;
-  cdpMarkets: ReadonlyArray<CdpBorrowingRevenueMarket>;
-  reserveYield: ReserveYieldResponse | null;
-  reserveDailySnapshots: ReadonlyArray<SusdsYieldDailySnapshotRow>;
-  reserveHistoryUnavailable?: boolean;
-  reserveHistoryFailed?: boolean;
-  reserveHistoryTruncated?: boolean;
-  reserveYieldFailed?: boolean;
-  swapFeesFailed?: boolean;
-  swapFeesApproximate?: boolean;
-  cdpDailySeriesFailed?: boolean;
-  cdpInputsApproximate?: boolean;
-  nowSeconds?: number;
-};
-
-type RevenueBucket = {
-  reserveYieldUsd: number;
-  swapFeesUsd: number;
-  cdpBorrowingUsd: number;
-};
-
-type ForecastSource = {
-  reserveDailyUsd: number | null;
-  reserve30dUsd: number | null;
-  reserve365dUsd: number | null;
-  swapDailyUsd: number | null;
-  cdpDailyUsd: number | null;
-  partialReasons: string[];
-};
-
-function currentDayBucket(nowSeconds: number): number {
-  return dayBucket(Math.floor(nowSeconds));
-}
-
-function dayBucket(timestampSeconds: number): number {
-  return Math.floor(timestampSeconds / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-}
-
-function isoDate(timestampSeconds: number): string {
-  return new Date(timestampSeconds * 1000).toLocaleDateString("en-US", {
-    month: "short",
-    day: "numeric",
-    year: "numeric",
-    timeZone: "UTC",
-  });
-}
-
-function emptyRevenueBucket(): RevenueBucket {
-  return { reserveYieldUsd: 0, swapFeesUsd: 0, cdpBorrowingUsd: 0 };
-}
-
-function addBucketValue(
-  buckets: Map<number, RevenueBucket>,
-  timestamp: number,
-  value: Partial<RevenueBucket>,
-): void {
-  if (timestamp < V3_REVENUE_LAUNCH_TIMESTAMP) return;
-  const bucketKey = dayBucket(timestamp);
-  if (bucketKey < V3_REVENUE_LAUNCH_TIMESTAMP) return;
-  const bucket = buckets.get(bucketKey) ?? emptyRevenueBucket();
-  bucket.reserveYieldUsd += value.reserveYieldUsd ?? 0;
-  bucket.swapFeesUsd += value.swapFeesUsd ?? 0;
-  bucket.cdpBorrowingUsd += value.cdpBorrowingUsd ?? 0;
-  buckets.set(bucketKey, bucket);
-}
-
-function numericUsdWei(value: string): number | null {
-  try {
-    const usd = weiToUsd(BigInt(value));
-    return Number.isFinite(usd) ? usd : null;
-  } catch {
-    return null;
-  }
-}
-
-function finiteOrNull(value: number | null | undefined): number | null {
-  return typeof value === "number" && Number.isFinite(value) ? value : null;
-}
-
-function buildRevenueBuckets(args: {
-  swapSeries: ReadonlyArray<ReturnType<typeof buildDailyFeeSeries>[number]>;
-  cdpDailySeries: ReadonlyArray<CdpBorrowingFeeSeriesPoint>;
-  reserveDailySnapshots: ReadonlyArray<SusdsYieldDailySnapshotRow>;
-}): Map<number, RevenueBucket> {
-  const buckets = new Map<number, RevenueBucket>();
-
-  for (const point of args.swapSeries) {
-    addBucketValue(buckets, point.timestamp, {
-      swapFeesUsd: point.protocolFeesUSD + point.lpFeesUSD,
-    });
-  }
-
-  for (const point of args.cdpDailySeries) {
-    addBucketValue(buckets, point.timestamp, {
-      cdpBorrowingUsd: point.totalFeesUSD,
-    });
-  }
-
-  const previousReserveTotalsBySource = new Map<string, number>();
-  const reserveRows = [...args.reserveDailySnapshots].sort(
-    (a, b) => Number(a.timestamp) - Number(b.timestamp),
-  );
-  for (const row of reserveRows) {
-    const timestamp = Number(row.timestamp);
-    if (!Number.isFinite(timestamp)) continue;
-    const totalYieldUsd = numericUsdWei(row.totalEarnedYieldUsdWei);
-    if (totalYieldUsd === null) continue;
-    const sourceKey = `${row.chainId}:${row.token.toLowerCase()}`;
-    const previousTotalUsd = previousReserveTotalsBySource.get(sourceKey);
-    const baselineUsd =
-      previousTotalUsd ?? reserveSnapshotBaselineUsd(row, totalYieldUsd);
-    if (baselineUsd === null) continue;
-    const dailyYieldUsd = totalYieldUsd - baselineUsd;
-    previousReserveTotalsBySource.set(sourceKey, totalYieldUsd);
-    addBucketValue(buckets, timestamp, { reserveYieldUsd: dailyYieldUsd });
-  }
-
-  return buckets;
-}
-
-function reserveSnapshotBaselineUsd(
-  row: SusdsYieldDailySnapshotRow,
-  totalYieldUsd: number,
-): number | null {
-  const dailyYieldUsd = numericUsdWei(row.dailyEarnedYieldUsdWei);
-  return dailyYieldUsd === null ? null : totalYieldUsd - dailyYieldUsd;
-}
-
-function buildDailySeries(
-  buckets: ReadonlyMap<number, RevenueBucket>,
-  nowSeconds: number,
-  actualAvailability: ActualRevenueAvailability,
-): CanonicalRevenueDailyPoint[] {
-  const today = currentDayBucket(nowSeconds);
-  const endBucket = today;
-  const series: CanonicalRevenueDailyPoint[] = [];
-  for (
-    let timestamp = V3_REVENUE_LAUNCH_TIMESTAMP;
-    timestamp <= endBucket;
-    timestamp += SECONDS_PER_DAY
-  ) {
-    const bucket = buckets.get(timestamp) ?? emptyRevenueBucket();
-    const reserveStale =
-      actualAvailability.reserveStaleAfter !== null &&
-      timestamp > actualAvailability.reserveStaleAfter;
-    const reserveYieldUsd =
-      actualAvailability.reserve && !reserveStale
-        ? bucket.reserveYieldUsd
-        : null;
-    const swapFeesUsd = actualAvailability.swap ? bucket.swapFeesUsd : null;
-    const cdpBorrowingUsd = actualAvailability.cdp
-      ? bucket.cdpBorrowingUsd
-      : null;
-    const values = [reserveYieldUsd, swapFeesUsd, cdpBorrowingUsd];
-    const availableRevenueUsd = values.reduce<number>(
-      (sum, value) => sum + (value ?? 0),
-      0,
-    );
-    const totalRevenueUsd =
-      reserveYieldUsd === null ||
-      swapFeesUsd === null ||
-      cdpBorrowingUsd === null
-        ? null
-        : availableRevenueUsd;
-    series.push({
-      timestamp,
-      reserveYieldUsd,
-      swapFeesUsd,
-      cdpBorrowingUsd,
-      totalRevenueUsd,
-      availableRevenueUsd,
-    });
-  }
-  return series;
-}
-
-function yearStartBucket(nowSeconds: number): number {
-  const now = new Date(nowSeconds * 1000);
-  return Math.floor(Date.UTC(now.getUTCFullYear(), 0, 1, 0, 0, 0, 0) / 1000);
-}
-
-function periodWindows(
-  nowSeconds: number,
-): Record<
-  RevenuePeriodKey,
-  Pick<CanonicalRevenuePeriod, "key" | "title" | "subtitle" | "from" | "to">
-> {
-  const today = currentDayBucket(nowSeconds);
-  const tomorrow = today + SECONDS_PER_DAY;
-  const ytdStart = Math.max(
-    yearStartBucket(nowSeconds),
-    V3_REVENUE_LAUNCH_TIMESTAMP,
-  );
-  return {
-    allTimeSinceV3: {
-      key: "allTimeSinceV3",
-      title: "Total Revenue",
-      subtitle: `Since ${V3_REVENUE_LAUNCH_LABEL}`,
-      from: V3_REVENUE_LAUNCH_TIMESTAMP,
-      to: tomorrow,
-    },
-    ytd: {
-      key: "ytd",
-      title: "Year To Date",
-      subtitle: `Calendar YTD since ${isoDate(ytdStart)}`,
-      from: ytdStart,
-      to: tomorrow,
-    },
-    last30d: {
-      key: "last30d",
-      title: "Last 30 Days",
-      subtitle: "Rolling UTC daily buckets",
-      from: today - 29 * SECONDS_PER_DAY,
-      to: tomorrow,
-    },
-    last7d: {
-      key: "last7d",
-      title: "Last 7 Days",
-      subtitle: "Rolling UTC daily buckets",
-      from: today - 6 * SECONDS_PER_DAY,
-      to: tomorrow,
-    },
-  };
-}
-
-function sumDailySeriesWindow(
-  series: ReadonlyArray<CanonicalRevenueDailyPoint>,
-  from: number,
-  to: number,
-): Pick<
+export {
+  V3_REVENUE_LAUNCH_LABEL,
+  V3_REVENUE_LAUNCH_TIMESTAMP,
+} from "./canonical-revenue/constants";
+export type {
+  BuildCanonicalRevenueArgs,
+  CanonicalRevenueDailyPoint,
+  CanonicalRevenueForecast,
   CanonicalRevenuePeriod,
-  "totalUsd" | "reserveYieldUsd" | "swapFeesUsd" | "cdpBorrowingUsd"
-> & { availableTotalUsd: number } {
-  let reserveYieldUsd = 0;
-  let swapFeesUsd = 0;
-  let cdpBorrowingUsd = 0;
-  let reserveYieldUnavailable = false;
-  let swapFeesUnavailable = false;
-  let cdpBorrowingUnavailable = false;
-  for (const point of series) {
-    if (point.timestamp < from || point.timestamp >= to) continue;
-    if (point.reserveYieldUsd === null) {
-      reserveYieldUnavailable = true;
-    } else {
-      reserveYieldUsd += point.reserveYieldUsd;
-    }
-    if (point.swapFeesUsd === null) {
-      swapFeesUnavailable = true;
-    } else {
-      swapFeesUsd += point.swapFeesUsd;
-    }
-    if (point.cdpBorrowingUsd === null) {
-      cdpBorrowingUnavailable = true;
-    } else {
-      cdpBorrowingUsd += point.cdpBorrowingUsd;
-    }
-  }
-  const availableTotalUsd = reserveYieldUsd + swapFeesUsd + cdpBorrowingUsd;
-  const totalUnavailable =
-    reserveYieldUnavailable || swapFeesUnavailable || cdpBorrowingUnavailable;
-  return {
-    reserveYieldUsd: reserveYieldUnavailable ? null : reserveYieldUsd,
-    swapFeesUsd: swapFeesUnavailable ? null : swapFeesUsd,
-    cdpBorrowingUsd: cdpBorrowingUnavailable ? null : cdpBorrowingUsd,
-    availableTotalUsd,
-    totalUsd: totalUnavailable ? null : availableTotalUsd,
-  };
-}
-
-function buildPeriods(
-  series: ReadonlyArray<CanonicalRevenueDailyPoint>,
-  nowSeconds: number,
-  partialReasons: string[],
-): Record<RevenuePeriodKey, CanonicalRevenuePeriod> {
-  const windows = periodWindows(nowSeconds);
-  return {
-    allTimeSinceV3: {
-      ...windows.allTimeSinceV3,
-      ...sumDailySeriesWindow(
-        series,
-        windows.allTimeSinceV3.from,
-        windows.allTimeSinceV3.to,
-      ),
-      partialReasons,
-    },
-    ytd: {
-      ...windows.ytd,
-      ...sumDailySeriesWindow(series, windows.ytd.from, windows.ytd.to),
-      partialReasons,
-    },
-    last30d: {
-      ...windows.last30d,
-      ...sumDailySeriesWindow(series, windows.last30d.from, windows.last30d.to),
-      partialReasons,
-    },
-    last7d: {
-      ...windows.last7d,
-      ...sumDailySeriesWindow(series, windows.last7d.from, windows.last7d.to),
-      partialReasons,
-    },
-  };
-}
-
-function completedWindowAverage<T>(args: {
-  points: ReadonlyArray<T>;
-  value: (point: T) => number;
-  timestamp: (point: T) => number;
-  nowSeconds: number;
-  trailingDays: number;
-  minimumBuckets: number;
-  minimumTimestamp?: number;
-}): { dailyAverageUsd: number | null; buckets: number } {
-  const today = currentDayBucket(args.nowSeconds);
-  const from = Math.max(
-    today - args.trailingDays * SECONDS_PER_DAY,
-    args.minimumTimestamp === undefined
-      ? Number.NEGATIVE_INFINITY
-      : dayBucket(args.minimumTimestamp),
-  );
-  const values = new Map<number, number>();
-  for (const point of args.points) {
-    const timestamp = args.timestamp(point);
-    const bucket = dayBucket(timestamp);
-    if (bucket < from || bucket >= today) continue;
-    values.set(bucket, (values.get(bucket) ?? 0) + args.value(point));
-  }
-  if (values.size < args.minimumBuckets) {
-    return { dailyAverageUsd: null, buckets: values.size };
-  }
-  let total = 0;
-  for (const value of values.values()) total += value;
-  return { dailyAverageUsd: total / values.size, buckets: values.size };
-}
-
-function marketProtocolShareBps(
-  market: CdpBorrowingRevenueMarket,
-): number | null {
-  if (market.spYieldSplitBps < 0 || market.spYieldSplitBps > 10_000) {
-    return null;
-  }
-  return 10_000 - market.spYieldSplitBps;
-}
-
-function computeCdpInterestDailyRunRate(
-  markets: ReadonlyArray<CdpBorrowingRevenueMarket>,
-): { dailyUsd: number | null; partialReason: string | null } {
-  let annualProtocolRunRateUsd = 0;
-  for (const market of markets) {
-    const protocolShareBps = marketProtocolShareBps(market);
-    if (protocolShareBps === null) {
-      return {
-        dailyUsd: null,
-        partialReason:
-          "CDP forecast unavailable: live protocol split is not indexed for every market.",
-      };
-    }
-    annualProtocolRunRateUsd +=
-      market.annualInterestRunRateUSD * (protocolShareBps / 10_000);
-  }
-  return { dailyUsd: annualProtocolRunRateUsd / 365, partialReason: null };
-}
-
-function reserveForecastPartialReasons(
-  reserveYield: ReserveYieldResponse,
-): string[] {
-  const reasons: string[] = [];
-  const unavailableSymbols = reserveYield.forecastUnavailableSymbols;
-  if (unavailableSymbols.length > 0) {
-    reasons.push(
-      `Reserve forecast excludes holdings without annual-rate sources: ${unavailableSymbols.join(", ")}.`,
-    );
-  }
-  if (reserveYield.holdingsError !== null) {
-    reasons.push(
-      "Reserve forecast partial: reserve holdings failed to load completely.",
-    );
-  }
-  if (reserveYield.rateError !== null) {
-    reasons.push(`Reserve forecast partial: ${reserveYield.rateError}`);
-  }
-  return reasons;
-}
-
-function buildForecastSource(args: {
-  reserveYield: ReserveYieldResponse | null;
-  swapSeries: ReadonlyArray<ReturnType<typeof buildDailyFeeSeries>[number]>;
-  swapFeesFailed: boolean;
-  swapFeesApproximate: boolean;
-  cdpDailySeries: ReadonlyArray<CdpBorrowingFeeSeriesPoint>;
-  cdpMarkets: ReadonlyArray<CdpBorrowingRevenueMarket>;
-  cdpDailySeriesFailed: boolean;
-  cdpInputsApproximate: boolean;
-  nowSeconds: number;
-}): ForecastSource {
-  const partialReasons: string[] = [];
-  const reserveDailyUsd = finiteOrNull(args.reserveYield?.dailyRunRateUsd);
-  const reserve30dUsd = finiteOrNull(args.reserveYield?.next30dUsd);
-  const reserve365dUsd = finiteOrNull(args.reserveYield?.next365dUsd);
-  if (args.reserveYield === null || reserveDailyUsd === null) {
-    partialReasons.push(
-      "Reserve forecast unavailable: current annual rate or reserve balances did not load.",
-    );
-  } else {
-    partialReasons.push(...reserveForecastPartialReasons(args.reserveYield));
-  }
-
-  const swapAverage = args.swapFeesFailed
-    ? { dailyAverageUsd: null, buckets: 0 }
-    : completedWindowAverage({
-        points: args.swapSeries,
-        value: (point) => point.protocolFeesUSD + point.lpFeesUSD,
-        timestamp: (point) => point.timestamp,
-        nowSeconds: args.nowSeconds,
-        trailingDays: 30,
-        minimumBuckets: 7,
-        minimumTimestamp: V3_REVENUE_LAUNCH_TIMESTAMP,
-      });
-  if (args.swapFeesFailed) {
-    partialReasons.push(
-      "Swap forecast unavailable: swap fee history failed to load.",
-    );
-  } else if (swapAverage.dailyAverageUsd === null) {
-    partialReasons.push(
-      `Swap forecast unavailable: only ${swapAverage.buckets} completed daily buckets loaded.`,
-    );
-  } else if (args.swapFeesApproximate) {
-    partialReasons.push(
-      "Swap forecast partial: swap fee history is approximate.",
-    );
-  }
-
-  let cdpDailyUsd: number | null = null;
-  if (args.cdpDailySeriesFailed) {
-    partialReasons.push(
-      "CDP forecast unavailable: borrowing revenue inputs failed to load.",
-    );
-  } else {
-    if (args.cdpInputsApproximate) {
-      partialReasons.push(
-        "CDP forecast partial: borrowing revenue inputs are approximate.",
-      );
-    }
-    const cdpInterest = computeCdpInterestDailyRunRate(args.cdpMarkets);
-    if (cdpInterest.partialReason !== null) {
-      partialReasons.push(cdpInterest.partialReason);
-    }
-    const cdpUpfrontAverage = completedWindowAverage({
-      points: args.cdpDailySeries,
-      value: (point) => point.upfrontFeesUSD,
-      timestamp: (point) => point.timestamp,
-      nowSeconds: args.nowSeconds,
-      trailingDays: 30,
-      minimumBuckets: 1,
-      minimumTimestamp: V3_REVENUE_LAUNCH_TIMESTAMP,
-    });
-    const cdpUpfrontDailyUsd = cdpUpfrontAverage.dailyAverageUsd ?? 0;
-    cdpDailyUsd =
-      cdpInterest.dailyUsd === null
-        ? null
-        : cdpInterest.dailyUsd + cdpUpfrontDailyUsd;
-  }
-
-  return {
-    reserveDailyUsd,
-    reserve30dUsd,
-    reserve365dUsd,
-    swapDailyUsd: swapAverage.dailyAverageUsd,
-    cdpDailyUsd,
-    partialReasons,
-  };
-}
-
-function forecastAssumption(days: number): string {
-  return [
-    "Forecast based on current reserve balances, trailing swap activity, and CDP current run-rate.",
-    `Reserve uses non-compounding forecast math: balance x annual rate x ${days} / 365.`,
-    "AUSD and stETH are forecast-only; sUSDS actuals come from indexed earned-yield snapshots.",
-    "Swap uses the last 30 completed daily fee buckets; CDP uses current protocol-share interest run-rate plus trailing upfront fees.",
-  ].join("\n");
-}
-
-function buildForecast(
-  source: ForecastSource,
-  key: RevenueForecastKey,
-  title: string,
-  subtitle: string,
-  days: number,
-): CanonicalRevenueForecast {
-  const reserveYieldUsd =
-    days === 30
-      ? source.reserve30dUsd
-      : days === 365
-        ? source.reserve365dUsd
-        : source.reserveDailyUsd === null
-          ? null
-          : source.reserveDailyUsd * days;
-  const swapFeesUsd =
-    source.swapDailyUsd === null ? null : source.swapDailyUsd * days;
-  const cdpBorrowingUsd =
-    source.cdpDailyUsd === null ? null : source.cdpDailyUsd * days;
-  const values = [reserveYieldUsd, swapFeesUsd, cdpBorrowingUsd].filter(
-    (value): value is number => value !== null,
-  );
-  return {
-    key,
-    title,
-    subtitle,
-    days,
-    totalUsd:
-      values.length === 0
-        ? null
-        : values.reduce((sum, value) => sum + value, 0),
-    reserveYieldUsd,
-    swapFeesUsd,
-    cdpBorrowingUsd,
-    partialReasons: source.partialReasons,
-    assumption: forecastAssumption(days),
-  };
-}
-
-function buildForecasts(
-  source: ForecastSource,
-): Record<RevenueForecastKey, CanonicalRevenueForecast> {
-  return {
-    next7d: buildForecast(source, "next7d", "7d Forecast", "Next 7 days", 7),
-    next30d: buildForecast(
-      source,
-      "next30d",
-      "Monthly Forecast",
-      "Next 30 days",
-      30,
-    ),
-    next365d: buildForecast(
-      source,
-      "next365d",
-      "Annual Forecast",
-      "Next 365 days",
-      365,
-    ),
-  };
-}
-
-function hasSusdsReserveYieldSignal(
-  reserveYield: ReserveYieldResponse | null,
-): boolean {
-  if (reserveYield === null) return false;
-  const earnedYieldUsd = finiteOrNull(reserveYield.earnedYieldUsd);
-  if (earnedYieldUsd !== null && earnedYieldUsd > 0) return true;
-  return reserveYield.holdings.some(
-    (holding) =>
-      holding.assetSymbol.toUpperCase() === "SUSDS" && holding.principalUsd > 0,
-  );
-}
-
-function latestReserveSnapshotBucket(
-  reserveDailySnapshots: ReadonlyArray<SusdsYieldDailySnapshotRow>,
-): number | null {
-  let latest: number | null = null;
-  for (const row of reserveDailySnapshots) {
-    const timestamp = Number(row.timestamp);
-    if (!Number.isFinite(timestamp)) continue;
-    const bucket = dayBucket(timestamp);
-    if (latest === null || bucket > latest) latest = bucket;
-  }
-  return latest;
-}
-
-function reserveStaleAfterBucket(
-  args: BuildCanonicalRevenueArgs,
-): number | null {
-  if (args.reserveHistoryFailed || args.reserveHistoryUnavailable) return null;
-  const latestBucket = latestReserveSnapshotBucket(args.reserveDailySnapshots);
-  if (latestBucket === null) return null;
-  return latestBucket <
-    currentDayBucket(args.nowSeconds ?? Math.floor(Date.now() / 1000))
-    ? latestBucket
-    : null;
-}
-
-function buildActualAvailability(
-  args: BuildCanonicalRevenueArgs,
-): ActualRevenueAvailability {
-  const reserveHistoryUnavailable =
-    args.reserveHistoryFailed === true ||
-    args.reserveHistoryUnavailable === true ||
-    (args.reserveDailySnapshots.length === 0 &&
-      (args.reserveYieldFailed === true ||
-        hasSusdsReserveYieldSignal(args.reserveYield)));
-  return {
-    reserve: !reserveHistoryUnavailable,
-    reserveStaleAfter: reserveStaleAfterBucket(args),
-    swap: args.swapFeesFailed !== true,
-    cdp: args.cdpDailySeriesFailed !== true,
-  };
-}
-
-function buildPartialReasons(args: BuildCanonicalRevenueArgs): string[] {
-  const reasons: string[] = [];
-  if (args.swapFeesFailed) reasons.push("Swap fee history failed to load.");
-  if (!args.swapFeesFailed && args.swapFeesApproximate) {
-    reasons.push("Swap fee history is approximate.");
-  }
-  if (args.cdpDailySeriesFailed) {
-    reasons.push("CDP borrowing revenue history failed to load.");
-  }
-  if (!args.cdpDailySeriesFailed && args.cdpInputsApproximate) {
-    reasons.push("CDP borrowing history is approximate.");
-  }
-  if (args.reserveHistoryFailed) {
-    reasons.push("Reserve earned-yield history failed to load.");
-  } else if (args.reserveHistoryUnavailable) {
-    reasons.push("Reserve earned-yield history is not indexed yet.");
-  } else if (
-    args.reserveYieldFailed &&
-    args.reserveDailySnapshots.length === 0
-  ) {
-    reasons.push(
-      "Reserve earned-yield actuals unavailable: current reserve yield failed to load before any snapshots were indexed.",
-    );
-  } else if (
-    args.reserveDailySnapshots.length === 0 &&
-    hasSusdsReserveYieldSignal(args.reserveYield)
-  ) {
-    reasons.push("Reserve earned-yield history has no sUSDS snapshots yet.");
-  } else {
-    const staleAfter = reserveStaleAfterBucket(args);
-    if (staleAfter !== null) {
-      reasons.push(
-        `Reserve earned-yield history is stale; latest snapshot is ${isoDate(staleAfter)}.`,
-      );
-    }
-  }
-  if (args.reserveHistoryTruncated) {
-    reasons.push("Reserve earned-yield history exceeded the pagination cap.");
-  }
-  return reasons;
-}
-
-function partialReasonsForStream(args: {
-  streamKey: CanonicalRevenueStream["key"];
-  reasons: readonly string[];
-}): string[] {
-  const needle =
-    args.streamKey === "cdp"
-      ? "cdp"
-      : args.streamKey === "swap"
-        ? "swap"
-        : "reserve";
-  return [
-    ...new Set(
-      args.reasons.filter((reason) => reason.toLowerCase().includes(needle)),
-    ),
-  ];
-}
-
-function forecastPartialReasonsForStream(args: {
-  streamKey: CanonicalRevenueStream["key"];
-  forecasts: Record<RevenueForecastKey, CanonicalRevenueForecast>;
-}): string[] {
-  return partialReasonsForStream({
-    streamKey: args.streamKey,
-    reasons: [
-      ...args.forecasts.next7d.partialReasons,
-      ...args.forecasts.next30d.partialReasons,
-      ...args.forecasts.next365d.partialReasons,
-    ],
-  });
-}
-
-function buildStreams(args: {
-  periods: Record<RevenuePeriodKey, CanonicalRevenuePeriod>;
-  forecasts: Record<RevenueForecastKey, CanonicalRevenueForecast>;
-  reserveYield: ReserveYieldResponse | null;
-  partialReasons: string[];
-}): Record<CanonicalRevenueStream["key"], CanonicalRevenueStream> {
-  const allPeriod = args.periods.allTimeSinceV3;
-  const reserveActualPartialReasons = partialReasonsForStream({
-    streamKey: "reserve",
-    reasons: args.partialReasons,
-  });
-  const reserveForecastPartialReasons = forecastPartialReasonsForStream({
-    streamKey: "reserve",
-    forecasts: args.forecasts,
-  });
-  const swapActualPartialReasons = partialReasonsForStream({
-    streamKey: "swap",
-    reasons: args.partialReasons,
-  });
-  const swapForecastPartialReasons = forecastPartialReasonsForStream({
-    streamKey: "swap",
-    forecasts: args.forecasts,
-  });
-  const cdpActualPartialReasons = partialReasonsForStream({
-    streamKey: "cdp",
-    reasons: args.partialReasons,
-  });
-  const cdpForecastPartialReasons = forecastPartialReasonsForStream({
-    streamKey: "cdp",
-    forecasts: args.forecasts,
-  });
-  return {
-    reserve: {
-      key: "reserve",
-      title: "Reserve Yield",
-      actualUsd: allPeriod.reserveYieldUsd,
-      forecast30dUsd: args.forecasts.next30d.reserveYieldUsd,
-      forecast365dUsd: args.forecasts.next365d.reserveYieldUsd,
-      subtitle:
-        args.reserveYield !== null
-          ? "sUSDS actual yield; AUSD and stETH forecast-only"
-          : "Reserve actuals from sUSDS snapshots",
-      actualPartialReasons: reserveActualPartialReasons,
-      forecastPartialReasons: reserveForecastPartialReasons,
-      partialReasons: [
-        ...new Set([
-          ...reserveActualPartialReasons,
-          ...reserveForecastPartialReasons,
-        ]),
-      ],
-    },
-    swap: {
-      key: "swap",
-      title: "Swap Fees",
-      actualUsd: allPeriod.swapFeesUsd,
-      forecast30dUsd: args.forecasts.next30d.swapFeesUsd,
-      forecast365dUsd: args.forecasts.next365d.swapFeesUsd,
-      subtitle: "Protocol fee snapshots across v3 pools",
-      actualPartialReasons: swapActualPartialReasons,
-      forecastPartialReasons: swapForecastPartialReasons,
-      partialReasons: [
-        ...new Set([
-          ...swapActualPartialReasons,
-          ...swapForecastPartialReasons,
-        ]),
-      ],
-    },
-    cdp: {
-      key: "cdp",
-      title: "CDP Borrowing Revenue",
-      actualUsd: allPeriod.cdpBorrowingUsd,
-      forecast30dUsd: args.forecasts.next30d.cdpBorrowingUsd,
-      forecast365dUsd: args.forecasts.next365d.cdpBorrowingUsd,
-      subtitle: "Protocol share of upfront fees and interest",
-      actualPartialReasons: cdpActualPartialReasons,
-      forecastPartialReasons: cdpForecastPartialReasons,
-      partialReasons: [
-        ...new Set([...cdpActualPartialReasons, ...cdpForecastPartialReasons]),
-      ],
-    },
-  };
-}
+  CanonicalRevenueResult,
+  CanonicalRevenueStream,
+  RevenueForecastKey,
+  RevenuePeriodKey,
+  SusdsYieldDailySnapshotRow,
+} from "./canonical-revenue/types";
 
 export function buildCanonicalRevenue({
   networkData,
@@ -881,7 +60,7 @@ export function buildCanonicalRevenue({
     cdpDailySeries,
     reserveDailySnapshots,
   });
-  const actualAvailability = buildActualAvailability({
+  const canonicalArgs = {
     networkData,
     cdpDailySeries,
     cdpMarkets,
@@ -896,24 +75,10 @@ export function buildCanonicalRevenue({
     cdpDailySeriesFailed,
     cdpInputsApproximate,
     nowSeconds,
-  });
+  };
+  const actualAvailability = buildActualAvailability(canonicalArgs);
   const dailySeries = buildDailySeries(buckets, nowSeconds, actualAvailability);
-  const partialReasons = buildPartialReasons({
-    networkData,
-    cdpDailySeries,
-    cdpMarkets,
-    reserveYield,
-    reserveDailySnapshots,
-    reserveHistoryUnavailable,
-    reserveHistoryFailed,
-    reserveHistoryTruncated,
-    reserveYieldFailed,
-    swapFeesFailed,
-    swapFeesApproximate,
-    cdpDailySeriesFailed,
-    cdpInputsApproximate,
-    nowSeconds,
-  });
+  const partialReasons = buildPartialReasons(canonicalArgs);
   const periods = buildPeriods(dailySeries, nowSeconds, partialReasons);
   const forecasts = buildForecasts(
     buildForecastSource({
