+++ ui-dashboard/src/lib/canonical-revenue/streams.ts
@@ -1,0 +1,122 @@
+import type {
+  CanonicalRevenueForecast,
+  CanonicalRevenuePeriod,
+  CanonicalRevenueStream,
+  RevenueForecastKey,
+  RevenuePeriodKey,
+} from "./types";
+import type { ReserveYieldResponse } from "@/lib/reserve-yield";
+
+function partialReasonsForStream(args: {
+  streamKey: CanonicalRevenueStream["key"];
+  reasons: readonly string[];
+}): string[] {
+  const needle =
+    args.streamKey === "cdp"
+      ? "cdp"
+      : args.streamKey === "swap"
+        ? "swap"
+        : "reserve";
+  return [
+    ...new Set(
+      args.reasons.filter((reason) => reason.toLowerCase().includes(needle)),
+    ),
+  ];
+}
+
+function forecastPartialReasonsForStream(args: {
+  streamKey: CanonicalRevenueStream["key"];
+  forecasts: Record<RevenueForecastKey, CanonicalRevenueForecast>;
+}): string[] {
+  return partialReasonsForStream({
+    streamKey: args.streamKey,
+    reasons: [
+      ...args.forecasts.next7d.partialReasons,
+      ...args.forecasts.next30d.partialReasons,
+      ...args.forecasts.next365d.partialReasons,
+    ],
+  });
+}
+
+export function buildStreams(args: {
+  periods: Record<RevenuePeriodKey, CanonicalRevenuePeriod>;
+  forecasts: Record<RevenueForecastKey, CanonicalRevenueForecast>;
+  reserveYield: ReserveYieldResponse | null;
+  partialReasons: string[];
+}): Record<CanonicalRevenueStream["key"], CanonicalRevenueStream> {
+  const allPeriod = args.periods.allTimeSinceV3;
+  const reserveActualPartialReasons = partialReasonsForStream({
+    streamKey: "reserve",
+    reasons: args.partialReasons,
+  });
+  const reserveForecastPartialReasons = forecastPartialReasonsForStream({
+    streamKey: "reserve",
+    forecasts: args.forecasts,
+  });
+  const swapActualPartialReasons = partialReasonsForStream({
+    streamKey: "swap",
+    reasons: args.partialReasons,
+  });
+  const swapForecastPartialReasons = forecastPartialReasonsForStream({
+    streamKey: "swap",
+    forecasts: args.forecasts,
+  });
+  const cdpActualPartialReasons = partialReasonsForStream({
+    streamKey: "cdp",
+    reasons: args.partialReasons,
+  });
+  const cdpForecastPartialReasons = forecastPartialReasonsForStream({
+    streamKey: "cdp",
+    forecasts: args.forecasts,
+  });
+  return {
+    reserve: {
+      key: "reserve",
+      title: "Reserve Yield",
+      actualUsd: allPeriod.reserveYieldUsd,
+      forecast30dUsd: args.forecasts.next30d.reserveYieldUsd,
+      forecast365dUsd: args.forecasts.next365d.reserveYieldUsd,
+      subtitle:
+        args.reserveYield !== null
+          ? "sUSDS actual yield; AUSD and stETH forecast-only"
+          : "Reserve actuals from sUSDS snapshots",
+      actualPartialReasons: reserveActualPartialReasons,
+      forecastPartialReasons: reserveForecastPartialReasons,
+      partialReasons: [
+        ...new Set([
+          ...reserveActualPartialReasons,
+          ...reserveForecastPartialReasons,
+        ]),
+      ],
+    },
+    swap: {
+      key: "swap",
+      title: "Swap Fees",
+      actualUsd: allPeriod.swapFeesUsd,
+      forecast30dUsd: args.forecasts.next30d.swapFeesUsd,
+      forecast365dUsd: args.forecasts.next365d.swapFeesUsd,
+      subtitle: "Protocol fee snapshots across v3 pools",
+      actualPartialReasons: swapActualPartialReasons,
+      forecastPartialReasons: swapForecastPartialReasons,
+      partialReasons: [
+        ...new Set([
+          ...swapActualPartialReasons,
+          ...swapForecastPartialReasons,
+        ]),
+      ],
+    },
+    cdp: {
+      key: "cdp",
+      title: "CDP Borrowing Revenue",
+      actualUsd: allPeriod.cdpBorrowingUsd,
+      forecast30dUsd: args.forecasts.next30d.cdpBorrowingUsd,
+      forecast365dUsd: args.forecasts.next365d.cdpBorrowingUsd,
+      subtitle: "Protocol share of upfront fees and interest",
+      actualPartialReasons: cdpActualPartialReasons,
+      forecastPartialReasons: cdpForecastPartialReasons,
+      partialReasons: [
+        ...new Set([...cdpActualPartialReasons, ...cdpForecastPartialReasons]),
+      ],
+    },
+  };
+}
